+++ android/src/main/AndroidManifest.xml
@@ -1,7 +1,4 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    package="com.example.flutterimagecompress">
-
-  <uses-sdk tools:overrideLibrary="androidx.heifwriter"/>
-
+<manifest xmlns:tools="http://schemas.android.com/tools"
+    package="com.fluttercandies.flutter_image_compress">
+    <uses-sdk tools:overrideLibrary="androidx.heifwriter"/>
 </manifest>
