--- android/src/main/kotlin/com/example/flutterimagecompress/exception/CompressError.kt
@@ -1,5 +1,0 @@
-package com.example.flutterimagecompress.exception
-
-import java.lang.Exception
-
-class CompressError(message: String) : Exception(message)
