--- android/src/main/kotlin/com/example/flutterimagecompress/format/FormatRegister.kt
@@ -1,19 +1,0 @@
-package com.example.flutterimagecompress.format
-
-import android.util.SparseArray
-import com.example.flutterimagecompress.handle.FormatHandler
-
-object FormatRegister {
-
-  private val formatMap = SparseArray<FormatHandler>()
-
-  fun registerFormat(handler: FormatHandler) {
-    formatMap.append(handler.type, handler)
-  }
-
-  fun findFormat(formatIndex: Int): FormatHandler? {
-    return formatMap.get(formatIndex)
-  }
-
-
-}
