--- android/src/main/kotlin/com/example/flutterimagecompress/util/TmpFileUtil.kt
@@ -1,14 +1,0 @@
-package com.example.flutterimagecompress.util
-
-import android.content.Context
-import java.io.File
-import java.util.*
-
-object TmpFileUtil {
-
-  fun createTmpFile(context: Context): File {
-    val string = UUID.randomUUID().toString()
-    return File(context.cacheDir, string)
-  }
-
-}
