+++ android/src/main/kotlin/com/fluttercandies/flutter_image_compress/exception/CompressError.kt
@@ -1,0 +1,5 @@
+package com.fluttercandies.flutter_image_compress.exception
+
+import java.lang.Exception
+
+class CompressError(message: String) : Exception(message)
