+++ android/src/main/kotlin/com/fluttercandies/flutter_image_compress/util/TmpFileUtil.kt
@@ -1,0 +1,12 @@
+package com.fluttercandies.flutter_image_compress.util
+
+import android.content.Context
+import java.io.File
+import java.util.*
+
+object TmpFileUtil {
+    fun createTmpFile(context: Context): File {
+        val string = UUID.randomUUID().toString()
+        return File(context.cacheDir, string)
+    }
+}
