+++ example/android/app/build.gradle
@@ -37,7 +37,7 @@ android {
     }
 
     defaultConfig {
-        applicationId "com.example.flutterimagecompressexample"
+        applicationId "com.fluttercandies.flutterimagecompressexample"
         minSdkVersion flutter.minSdkVersion
         targetSdkVersion flutter.targetSdkVersion
         versionCode flutterVersionCode.toInteger()
