--- example/android/app/src/main/gen/com/example/flutterimagecompressexample/BuildConfig.java
@@ -1,8 +1,0 @@
-/*___Generated_by_IDEA___*/
-
-package com.example.flutterimagecompressexample;
-
-/* This stub is only used by the IDE. It is NOT the BuildConfig class actually packed into the APK */
-public final class BuildConfig {
-  public final static boolean DEBUG = Boolean.parseBoolean(null);
-}
