--- example/android/app/src/main/gen/com/example/flutterimagecompressexample/Manifest.java
@@ -1,7 +1,0 @@
-/*___Generated_by_IDEA___*/
-
-package com.example.flutterimagecompressexample;
-
-/* This stub is only used by the IDE. It is NOT the Manifest class actually packed into the APK */
-public final class Manifest {
-}
