--- example/android/app/src/main/gen/com/example/flutterimagecompressexample/R.java
@@ -1,7 +1,0 @@
-/*___Generated_by_IDEA___*/
-
-package com.example.flutterimagecompressexample;
-
-/* This stub is only used by the IDE. It is NOT the R class actually packed into the APK */
-public final class R {
-}
