--- example/android/app/src/main/kotlin/com/example/flutterimagecompressexample/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.example.flutterimagecompressexample
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity() : FlutterActivity() {
-}
