+++ example/android/app/src/main/kotlin/com/fluttercandies/flutterimagecompressexample/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.fluttercandies.flutterimagecompressexample
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity() : FlutterActivity() {
+}
