+++ example/go/go.mod
@@ -8,9 +8,9 @@ require (
 	github.com/go-flutter-desktop/go-flutter v0.40.1
 	github.com/go-flutter-desktop/plugins/path_provider v0.4.0
 	github.com/nfnt/resize v0.0.0-20180221191011-83c6a9932646 // indirect
-	github.com/openflutter/flutter_image_compress/go v0.0.0-00010101000000-000000000000
+	github.com/fluttercandies/flutter_image_compress/go v0.0.0-00010101000000-000000000000
 	github.com/pkg/errors v0.9.1
 	golang.org/x/image v0.0.0-20200430140353-33d19683fad8 // indirect
 )
 
-replace github.com/openflutter/flutter_image_compress/go => /Users/caijinglong/Documents/GitHub/flutter_image_compress/go
+replace github.com/fluttercandies/flutter_image_compress/go => /Users/caijinglong/Documents/GitHub/flutter_image_compress/go
