+++ example/ios/Runner/Info.plist
@@ -11,7 +11,9 @@
 	<key>CFBundleInfoDictionaryVersion</key>
 	<string>6.0</string>
 	<key>CFBundleName</key>
-	<string>flutter_image_compress_example</string>
+	<string>image_compress_example</string>
+	<key>CFBundleDisplayName</key>
+	<string>image_compress Example</string>
 	<key>CFBundlePackageType</key>
 	<string>APPL</string>
 	<key>CFBundleShortVersionString</key>
