--- ios/Classes/FlutterImageCompressPlugin.h
@@ -1,7 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface FlutterImageCompressPlugin : NSObject<FlutterPlugin>
-
-+(BOOL) showLog;
-
-@end
