+++ ios/Classes/ImageCompressPlugin.h
@@ -1,0 +1,7 @@
+#import <Flutter/Flutter.h>
+
+@interface ImageCompressPlugin : NSObject<FlutterPlugin>
+
++(BOOL) showLog;
+
+@end
