+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: flutter_image_compress
 description: Compress image with native (Objective-C/Kotlin) with faster speed. Support Android/iOS.
-repository: https://github.com/OpenFlutter/flutter_image_compress
+repository: https://github.com/fluttercandies/flutter_image_compress
 version: 1.1.1
 
 environment:
@@ -15,7 +15,7 @@ flutter:
   plugin:
     platforms:
       android:
-        package: com.example.flutterimagecompress
-        pluginClass: FlutterImageCompressPlugin
+        package: com.fluttercandies.flutter_image_compress
+        pluginClass: ImageCompressPlugin
       ios:
-        pluginClass: FlutterImageCompressPlugin
+        pluginClass: ImageCompressPlugin
